requirements.txt
albumentations==1.4.21
diffusers==0.31.0
numpy==2.1.3
Pillow==11.0.0
scikit_learn==1.5.2
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.0
typed_argument_parser==1.10.1
wandb==0.18.7
